requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pydantic
python-dotenv
fastapi
reportlab